+++ src/decoder/utf8-decode.ts
@@ -3,14 +3,9 @@
  * Inspired by https://appspector.com/blog/how-to-improve-messagepack-javascript-parsing-speed-by-2-6-times
  * Follow instruction by https://zh.wikipedia.org/wiki/UTF-8#UTF-8%E7%9A%84%E7%B7%A8%E7%A2%BC%E6%96%B9%E5%BC%8F
  */
-export function utf8Decode(
-  bytes: Uint8Array,
-  offset = 0,
-  end?: number
-): string {
-  if (!end) {
-    end = bytes.byteLength
-  }
+export function utf8Decode(bytes: Uint8Array): string {
+  let offset = 0
+  const end = bytes.byteLength
   const out: Array<number> = []
   while (offset < end) {
     const firstByte = bytes[offset++]